refactor: check for absolute paths in the auxiliary vector
getexecname(3) on Solaris says it may return a relative path, and
testing indicates that getauxval(AT_EXECFN) may also return a relative
path on Linux. (elf_aux_info(AT_EXECPATH) on FreeBSD doesn't seem to do
this, but let's be cautious.)

This check is necessary because trying to exec() relative paths like
this may cause bugs. Fortunately, all of the affected OSes have fallback
methods that can be employed instead, so this shouldn't break anything.